Fixes compilation for when the rule is in a different place than the targets.
Originally whey they are in a different place, such as when compiling from another workspace, the
paths given to the compiler for output become inconsistent with the paths expected for import.
Furthermore, even if in the same workspace but at a different location, the compilation would fail
to put the files in the correct place.

This commit addresses that by allowing users to specify in the proto path things like
`"@other_repo//path/to:file.proto"` and have it be properly handled. Since this moves files around,
there is a possibility of collision (multiple rules creating the same file). Although bazel handles
that for the user, the message can be a little bit cryptic.